feat(classes): Detect property methods of classes
When functions/methods are decorated with `@property`, it is detected
and such methods are described as "property methods" in the output. The
detection is performed in `FunctionsExtractor`. Currently, only the
vanilla `@property` is detected. Getter/setter variants, like
`@property.getter`, are not.

Additionally, removed functions/methods are now detected.

Implementation notes:

- `pyff_function` was made to operate on `FunctionSummary` instances,
  just like `pyff_class` started to do. To achieve this,
  `FunctionsExtractor` was made to save the AST node to
  `FunctionSummary` instances.