feat(attributes): Detect differences in class attributes
pyff is now able to detect and compare attributes used in methods.
Attribute is considered to be any name that is used as an attribute of
'self' and is assigned into in some method.

To accomodate this, `ClassSummary` was changed to track methods and
attributes as sets of names (previously, only counts were tracked for
methods). This also made possible to simplify the `kitchensink.hl()`
method, because it no longer needs to handle integer arguments.